+++ app/src/main/java/com/example/playlistmaker/domain/api/PlayerRepository.kt
@@ -5,6 +5,6 @@ interface PlayerRepository {
     fun startPlayer()
     fun pausePlayer()
     fun releasePlayer()
-    fun getCurrentPosition(): Int
+    fun getCurrentPosition(): Long
     fun isPlaying(): Boolean
 }
